backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
sqlite3
pydantic==2.5.0
python-multipart==0.0.6
crewai==0.1.55
google-cloud-aiplatform==1.38.1
serpapi==0.1.5
reportlab==4.0.7
weasyprint==60.2
jinja2==3.1.2
httpx==0.25.2
python-dotenv==1.0.0